modern-home

+++ client/src/app/reservation-summary/reservation-summary.component.ts
@@ -1,7 +1,7 @@
 import { ReservationService } from './../reservation.service';
 import { Observable } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Reservation } from '../models/reservation';
 import { NavigationService } from '../navigation.service';
 
@@ -10,28 +10,29 @@ import { NavigationService } from '../navigation.service';
   templateUrl: './reservation-summary.component.html',
   styleUrls: ['./reservation-summary.component.css']
 })
-export class ReservationSummaryComponent {
-  reservation$: Observable<Reservation>; 
-  reservationId: string;
+export class ReservationSummaryComponent implements OnInit {
+  reservation$!: Observable<Reservation>;
+  reservationId!: string;
 
   constructor(
-    private route: ActivatedRoute, 
-    private reservationS: ReservationService, 
-    private navigation: NavigationService) { 
-      
-     this.reservationId = this.route.snapshot.paramMap.get('id') || "";
-     this.reservation$ = this.reservationS.getReservation(this.reservationId);
+    private route: ActivatedRoute,
+    private reservationS: ReservationService,
+    private navigationS: NavigationService) { }
+
+  ngOnInit() {
+    this.reservationId = this.route.snapshot.paramMap.get('id') || "";
+    this.reservation$ = this.reservationS.getReservation(this.reservationId);
   }
 
-
-  back(){
-    this.navigation.back();
+  back() {
+    this.navigationS.back();
   }
 
   delete() {
-    alert("Are You sure You want to cancel this reservation?");
+    if (!confirm("Are You sure You want to cancel this reservation?")) return;
+
     this.reservationS.deleteReservation(this.reservationId).subscribe();
-    this.navigation.back();
+    this.navigationS.back();
   }
 
 }
